pyproject.toml
[build-system]
requires = ["hatchling"]
build-backend = "hatchling.build"

[project]
name = "easy-docker-deploy"
version = "0.1.0"
description = "A command-line tool to easily deploy self-hosted applications using Docker"
readme = "README.md"
requires-python = ">=3.8"
license = "MIT"
keywords = ["docker", "deployment", "cli", "self-hosted"]
authors = [
    { name = "Your Name", email = "your.email@example.com" }
]
classifiers = [
    "Development Status :: 4 - Beta",
    "Environment :: Console",
    "Intended Audience :: Developers",
    "Intended Audience :: System Administrators",
    "License :: OSI Approved :: MIT License",
    "Operating System :: OS Independent",
    "Programming Language :: Python :: 3",
    "Programming Language :: Python :: 3.8",
    "Programming Language :: Python :: 3.9",
    "Programming Language :: Python :: 3.10",
    "Programming Language :: Python :: 3.11",
    "Topic :: System :: Installation/Setup",
    "Topic :: System :: Systems Administration",
    "Topic :: Utilities",
]
dependencies = [
    "typer>=0.9.0",
    "rich>=13.0.0",
    "pyyaml>=6.0",
    "requests>=2.28.0",
    "docker>=6.0.0",
]

[project.optional-dependencies]
dev = [
    "pytest>=7.0.0",
    "pytest-cov>=4.0.0",
    "mypy>=1.0.0",
    "ruff>=0.0.270",
    "black>=23.0.0",
]

[project.urls]
Homepage = "https://github.com/yourusername/easy-docker-deploy"
Documentation = "https://github.com/yourusername/easy-docker-deploy#readme"
Repository = "https://github.com/yourusername/easy-docker-deploy.git"
"Bug Tracker" = "https://github.com/yourusername/easy-docker-deploy/issues"

[project.scripts]
easy-docker-deploy = "easy_docker_deploy.cli.main:app"

[tool.ruff]
# Enable pycodestyle ('E'), pyflakes ('F'), and isort ('I') rules
select = ["E", "F", "I"]
ignore = []

# Same as Black.
line-length = 88

# Assume Python 3.8
target-version = "py38"

[tool.ruff.per-file-ignores]
"__init__.py" = ["F401"]

[tool.ruff.isort]
known-first-party = ["easy_docker_deploy"]

[tool.mypy]
python_version = "3.8"
warn_return_any = true
warn_unused_configs = true
disallow_untyped_defs = true
disallow_incomplete_defs = true
check_untyped_defs = true
disallow_untyped_decorators = true
no_implicit_optional = true
warn_redundant_casts = true
warn_unused_ignores = true
warn_no_return = true
warn_unreachable = true

[[tool.mypy.overrides]]
module = "tests.*"
disallow_untyped_defs = false
disallow_incomplete_defs = false

[tool.pytest.ini_options]
minversion = "7.0"
addopts = "-ra -q --cov=src/easy_docker_deploy --cov-report=term-missing"
testpaths = [
    "tests",
] 